refactor: extract LiveKeyboardOverlayController into focused extensions
Split 1,360-line controller into 7 files by responsibility:
- Core (753 lines): properties, init, window lifecycle, NSWindowDelegate
- +LayerState (139): layer change observation, one-shot override
- +LauncherSession (89): launcher activation/deactivation
- +KeyClickHandling (100): key click dispatch, keymap changes
- +AppSuppression (24): per-app hide/restore
- +Observers (218): notification observers (mapper, wizard, health)
- +Inspector (411): existing, unchanged
